serverPY/aisdb-api/requirements.txt
-e git+https://github.com/AISViz/AISdb.git@627276d95eb34639e0597f5feb585c197251479e#egg=aisdb
annotated-types==0.7.0
anyio==4.6.2.post1
attrs==24.2.0
beautifulsoup4==4.12.3
blinker==1.9.0
Brotli==1.1.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
fastapi==0.115.5
Flask==3.1.0
h11==0.14.0
httptools==0.6.4
idna==3.10
inflate64==1.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.2
maturin==1.7.6
multivolumefile==0.2.3
numpy==2.1.3
orjson==3.10.12
outcome==1.3.0.post0
packaging==24.2
pillow==11.0.0
psutil==6.1.0
psycopg==3.2.3
psycopg-binary==3.2.3
py7zr==0.22.0
pybcj==1.0.2
pycryptodomex==3.21.0
pydantic==2.10.2
pydantic_core==2.27.1
pyppmd==1.1.0
pyproj==3.7.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
PyYAML==6.0.2
pyzstd==0.16.2
requests==2.32.3
scipy==1.14.1
selenium==4.27.1
setuptools==75.6.0
shapely==2.0.6
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
starlette==0.41.3
texttable==1.7.0
toml==0.10.2
tqdm==4.67.1
trio==0.27.0
trio-websocket==0.11.1
typing_extensions==4.12.2
urllib3==2.2.3
uvicorn==0.32.1
uvloop==0.21.0
watchfiles==1.0.0
webdriver-manager==4.0.2
websocket-client==1.8.0
websockets==14.1
Werkzeug==3.1.3
wheel==0.45.1
wsproto==1.2.0
